+++ src/plugins/saved_objects_management/public/management_section/objects_table/components/flyout.test.tsx
@@ -221,6 +221,7 @@ describe('Flyout', () => {
           createNewCopies: true,
           overwrite: true,
         },
+        undefined,
         undefined
       );
       expect(component.state()).toMatchObject({