+++ CHANGELOG.md
@@ -1,7 +1,31 @@
 # CHANGELOG
 Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased](https://github.com/opensearch-project/OpenSearch-Dashboards/compare/1.3.17...1.3)
+## [Unreleased](https://github.com/opensearch-project/OpenSearch-Dashboards/compare/1.3.18...1.3)
+
+### 💥 Breaking Changes
+
+### Deprecations
+
+### 🛡   Security
+
+### 📈 Features/Enhancements
+
+### 🐛 Bug Fixes
+
+### 🚞 Infrastructure
+
+### 📝 Documentation
+
+### 🛠   Maintenance
+
+- [Version] Increment version to 1.3.19 ([#7733](https://github.com/opensearch-project/OpenSearch-Dashboards/pull/7733))
+
+### 🪛  Refactoring
+
+### 🔩 Tests
+
+## [1.3.18 - 2024-07-16](https://github.com/opensearch-project/OpenSearch-Dashboards/releases/tag/1.3.18)
 
 ### 💥 Breaking Changes